Moved invocation of sf-brag-to-diamonds, so it's right at the end of the postbuild (apart from publishing zipped logs).
This avoids the possibility of a build abort when parallel threaqds are trying to a) convert a log file into a brag file, and b) merge all brag files ready to determine the brag status.

It also makes more sense - don't calculate or send the BRAGG status until the build is on the cusp of completion.